requirements.txt
altgraph==0.17.2
anyio==3.6.1
beautifulsoup4==4.11.1
certifi==2022.6.15
future==0.18.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
pefile==2022.5.30
Pillow==9.2.0
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
PySide6==6.3.1
PySide6-Addons==6.3.1
PySide6-Essentials==6.3.1
pywin32-ctypes==0.2.0
rfc3986==1.5.0
shiboken6==6.3.1
sniffio==1.2.0
soupsieve==2.3.2.post1
